new skeleton for implementing paging
