+++ src/page/todo/index.js
@@ -1,0 +1,1 @@
+export { Todo } from './Todo'
